first draft of authentication plus rate limiting still need to test.

+++ api/businesses.js
@@ -2,15 +2,52 @@ const mongoose = require('mongoose');
 const express = require('express');
 const businessSchema = require('../models/business');
 const router = require('express').Router();
-// const businesses = require('../db/businesses');
-// const { reviews } = require('./reviews');
-// const { photos } = require('./photos');
-// const { validateAgainstSchema, extractValidFields } = require('../val/validation');
-
+const jwt = require('jsonwebtoken');
+const jwtMiddleware = require('../jwtMiddleware');
+const rateLimit = require('express-rate-limit');
 exports.router = router;
-// exports.businesses = businesses;
 
-router.post('/post', async (req, res) => {
+const unAuthenticatedLimiter = rateLimit({
+  windowMs: 60 * 1000,
+  max: 5,
+  message: 'Too many requests, please try again later.'
+});
+
+const authenticatedLimiter = rateLimit({
+  windowMs: 60 * 1000,
+  max: 10,
+  keyGenerator: (req) => req.user.id,
+  message: 'Too many requests, please try again later.'
+});
+
+
+
+router.get('/', unAuthenticatedLimiter, async (req, res) => {
+  try{
+    const data = await businessSchema.find();
+    res.json(data)
+  }
+  catch(error){
+      res.status(500).json({message: error.message})
+  }
+});
+
+router.get('/:id', unAuthenticatedLimiter, async (req, res) => {
+  try{
+    const business = await businessSchema.findById(req.params.id);
+    res.json(business)
+  }
+  catch(error){
+      res.status(500).json({message: error.message})
+  }
+});
+
+router.use(authenticatedLimiter);
+
+router.post('/post', jwtMiddleware, async (req, res) => {
+  if (req.body.userid !== req.user.id && !req.user.admin) {
+    return res.status(403).json({ message: 'Forbidden' });
+  }
   const business = new businessSchema({
     _id: new mongoose.Types.ObjectId(),
     id: req.body.id,
@@ -27,32 +64,17 @@ router.post('/post', async (req, res) => {
     email: req.body.email
   });
   try{
-    const savedBusiness = business.save();
+    const savedBusiness = await business.save();
     res.status(200).json(savedBusiness);
   } catch (err) {
     res.status(400).json({ message: err });
   }
 });
-router.get('/', async (req, res) => {
-  try{
-    const data = await businessSchema.find();
-    res.json(data)
-  }
-  catch(error){
-      res.status(500).json({message: error.message})
-  }
-});
 
-router.get('/:id', async (req, res) => {
-  try{
-    const business = await businessSchema.findById(req.params.id);
-    res.json(business)
+router.patch('/update/:id',jwtMiddleware, async (req, res) => {
+  if (req.body.userid !== req.user.id && !req.user.admin) {
+    return res.status(403).json({ message: 'Forbidden' });
   }
-  catch(error){
-      res.status(500).json({message: error.message})
-  }
-});
-router.patch('/update/:id', async (req, res) => {
   try {
     const id = req.params.id;
     const updatedData = req.body;
@@ -68,7 +90,10 @@ router.patch('/update/:id', async (req, res) => {
       res.status(400).json({ message: error.message })
   }
 });
-router.delete('/delete/:id', async(req, res) => {
+router.delete('/delete/:id', jwtMiddleware, async(req, res) => {
+  if (req.body.userid !== req.user.id && !req.user.admin) {
+    return res.status(403).json({ message: 'Forbidden' });
+  }
   try {
     const id = req.params.id;
     const data = await businessSchema.findByIdAndDelete(id)
