+++ api/photos.js
@@ -1,15 +1,51 @@
 const express = require('express');
 const mongoose = require('mongoose');
 const router = require('express').Router();
-// const { validateAgainstSchema, extractValidFields } = require('../val/validation');
+
 const photoSchema = require('../models/photo');
-// const photos = require('../db/photos');
-
+const jwtMiddleware = require('../jwtMiddleware');
+const rateLimit = require('express-rate-limit');
 exports.router = router;
-// exports.photos = photos;
 
+const unAuthenticatedLimiter = rateLimit({
+  windowMs: 60 * 1000,
+  max: 5,
+  message: 'Too many requests, please try again later.'
+});
 
-router.post('/post', async (req, res) => {
+const authenticatedLimiter = rateLimit({
+  windowMs: 60 * 1000,
+  max: 10,
+  keyGenerator: (req) => req.user.id,
+  message: 'Too many requests, please try again later.'
+});
+
+router.get('/', unAuthenticatedLimiter, async (req, res) => {
+  try{
+    const data = await photoSchema.find();
+    res.json(data)
+  }
+  catch(error){
+      res.status(500).json({message: error.message})
+  }
+});
+
+router.get('/:id', unAuthenticatedLimiter, async (req, res) => {
+  try{
+    const business = await photoSchema.findById(req.params.id);
+    res.json(business)
+  }
+  catch(error){
+      res.status(500).json({message: error.message})
+  }
+});
+
+router.use(authenticatedLimiter);
+router.post('/post', jwtMiddleware, async (req, res) => {
+  if (req.body.userid !== req.user.id && !req.user.admin) {
+    return res.status(403).json({ message: 'Forbidden' });
+  }
+  
   const photo = new photoSchema({
     _id: new mongoose.Types.ObjectId(),
     id: req.body.id,
@@ -24,26 +60,11 @@ router.post('/post', async (req, res) => {
     res.status(400).json({ message: err });
   }
 });
-router.get('/', async (req, res) => {
-  try{
-    const data = await photoSchema.find();
-    res.json(data)
-  }
-  catch(error){
-      res.status(500).json({message: error.message})
-  }
-});
 
-router.get('/:id', async (req, res) => {
-  try{
-    const business = await photoSchema.findById(req.params.id);
-    res.json(business)
+router.patch('/update/:id', jwtMiddleware, async (req, res) => {
+  if (req.body.userid !== req.user.id && !req.user.admin) {
+    return res.status(403).json({ message: 'Forbidden' });
   }
-  catch(error){
-      res.status(500).json({message: error.message})
-  }
-});
-router.patch('/update/:id', async (req, res) => {
   try {
     const id = req.params.id;
     const updatedData = req.body;
@@ -57,7 +78,10 @@ router.patch('/update/:id', async (req, res) => {
       res.status(400).json({ message: error.message })
   }
 });
-router.delete('/delete/:id', async(req, res) => {
+router.delete('/delete/:id', jwtMiddleware, async(req, res) => {
+  if (req.body.userid !== req.user.id && !req.user.admin) {
+    return res.status(403).json({ message: 'Forbidden' });
+  }
   try {
     const id = req.params.id;
     const data = await photoSchema.findByIdAndDelete(id)
