+++ api/reviews.js
@@ -1,14 +1,50 @@
 const mongoose = require('mongoose');
 const express = require('express');
 const router = require('express').Router();
-// const { validateAgainstSchema, extractValidFields } = require('../val/validation');
+const jwtMiddleware = require('../jwtMiddleware');
 const reviewSchema = require('../models/review');
-// const reviews = require('../db/reviews');
-
+const rateLimit = require('express-rate-limit');
 exports.router = router;
-// exports.reviews = reviews;
 
-router.post('/post', async (req, res) => {
+const unAuthenticatedLimiter = rateLimit({
+  windowMs: 60 * 1000,
+  max: 5,
+  message: 'Too many requests, please try again later.'
+});
+
+const authenticatedLimiter = rateLimit({
+  windowMs: 60 * 1000,
+  max: 10,
+  keyGenerator: (req) => req.user.id,
+  message: 'Too many requests, please try again later.'
+});
+
+router.get('/', unAuthenticatedLimiter, async (req, res) => {
+  try{
+    const data = await reviewSchema.find();
+    res.json(data)
+  }
+  catch(error){
+      res.status(500).json({message: error.message})
+  }
+});
+
+router.get('/:id', unAuthenticatedLimiter, async (req, res) => {
+  try{
+    const review = await reviewSchema.findById(req.params.id);
+    res.json(review)
+  }
+  catch(error){
+      res.status(500).json({message: error.message})
+  }
+});
+
+router.use(authenticatedLimiter);
+
+router.post('/post', jwtMiddleware, async (req, res) => {
+  if (req.body.userid !== req.user.id && !req.user.admin) {
+    return res.status(403).json({ message: 'Forbidden' });
+  }
   const review = new reviewSchema({
     _id: new mongoose.Types.ObjectId(),
     id: req.body.id,
@@ -26,26 +62,11 @@ router.post('/post', async (req, res) => {
   }
 });
 
-router.get('/', async (req, res) => {
-  try{
-    const data = await reviewSchema.find();
-    res.json(data)
-  }
-  catch(error){
-      res.status(500).json({message: error.message})
-  }
-});
 
-router.get('/:id', async (req, res) => {
-  try{
-    const review = await reviewSchema.findById(req.params.id);
-    res.json(review)
+router.patch('/update/:id', jwtMiddleware, async (req, res) => {
+  if (req.body.userid !== req.user.id && !req.user.admin) {
+    return res.status(403).json({ message: 'Forbidden' });
   }
-  catch(error){
-      res.status(500).json({message: error.message})
-  }
-});
-router.patch('/update/:id', async (req, res) => {
   try {
     const id = req.params.id;
     const updatedData = req.body;
@@ -59,7 +80,10 @@ router.patch('/update/:id', async (req, res) => {
       res.status(400).json({ message: error.message })
   }
 });
-router.delete('/delete/:id', async(req, res) => {
+router.delete('/delete/:id', jwtMiddleware, async(req, res) => {
+  if (req.body.userid !== req.user.id && !req.user.admin) {
+    return res.status(403).json({ message: 'Forbidden' });
+  }
   try {
     const id = req.params.id;
     const data = await reviewSchema.findByIdAndDelete(id)
